Update Github repo URL to not have a '.go' in it
I now know a bunch more about good repo naming for Go projects, and one
bit is that a '.go' does not normally belong in the name of the repo.
So I have changed the Github repo URL to not have that. This requires
updates to places where the repo URL is mentioned.