Assume use of TLS by default for known TLS ports
If you say just 'call host port', call normally defaults to connecting
with 'tcp'. If the port is on a shortlist of known TLS-using ports, we
now assume the protocol is 'tls' instead. This is more convenient, as I
keep doing things like 'call host https' and then remembering that I
have to say 'call tls host https' instead.

The implementation is a bit sleazy and some of our help text should
really be changed to match (but I can't think of a good short phrasing
for what we now do).