Allow specifying a timeout for making connections
The timeout is specified as a duration, which means one must append 's'
to get seconds. This is a bit unfortunate and user-unfriendly, but I
don't feel like trying much harder to do a little better here.

In the process, dial() was refactored so that both -b and -t apply to
TLS connections as well as non-TLS ones. This is the way it should have
been from the beginning, and the code is simpler now.