Add a -L option to append newlines to datagrams
When receiving datagram records, it is convenient to be able to append
newlines to datagrams that don't have them. In specific, this allows
call to be used to receive syslog datagrams, which are otherwise ASCII
but which don't have trailing newlines. Chris had a use for this
recently.

Possibly this should be supported outside of -l mode, but we'll see.