Switch option parsing to using github.com/pborman/getopt/v2
This adds long versions of most options, as well as supporting
standard Unix getopt style options parsing.

I'm done with the limitations and irritations of the standard
library's 'flag' package. I believe that all old flags are still
valid, but I'm not sure and I didn't keep careful track. All of the
major flags that I use are still there.